Remove `levenshteinDistanceWithString...`
Looks like it was causing crashes like this;
> xctest(8240,0x7fff77890310) malloc: *** error for object 0x100823c08:
> incorrect checksum for freed object - object was probably modified
> after being freed.
> *** set a breakpoint in malloc_error_break to debug

I didn't get what actually was wrong there so as possible solution I'm
proposing to remove it. Probably it was added because on different OS
(OS X and iOS) date formatter may return different values. But right now
the project configured that way that it runs tests only in OS X. And I
expect it returns the same values all the time.

Fixes #34